refactor: e2e cases
