+++ packages/tests/src/e2e/archive/teamsAgent/DebugCustomCopilotAgentNewBotForPython.tests.bak
@@ -12,13 +12,14 @@ import * as path from "path";
 
 import { it } from "@microsoft/extra-shot-mocha";
 
-import { CliHelper } from "../../commonlib/cliHelper";
+import { CliHelper } from "../../../commonlib/cliHelper";
+import { execAsync } from "../../../utils/commonUtils";
 import {
   cleanUpLocalProject,
   getTestFolder,
   getUniqueAppName,
   readContextMultiEnvV3,
-} from "../commonUtils";
+} from "../../commonUtils";
 import {
   deleteAadAppByClientId,
   deleteBot,
@@ -26,8 +27,7 @@ import {
   getAadAppByClientId,
   getBot,
   getTeamsApp,
-} from "./utility";
-import { execAsync } from "../../utils/commonUtils";
+} from "../../debug/utility";
 
 describe("Debug V3 command-and-response template", () => {
   const testFolder = getTestFolder();
@@ -49,16 +49,17 @@ describe("Debug V3 command-and-response template", () => {
   });
 
   it(
-    "OpenAI happy path: provision and deploy",
-    { testPlanCaseId: 28165244, author: "frankqian@microsoft.com" },
+    "Azure OpenAI happy path: provision and deploy",
+    { testPlanCaseId: 28957868, author: "frankqian@microsoft.com" },
     async function () {
       // create
       const myRecordAzOpenAI: Record<string, string> = {};
       myRecordAzOpenAI["programming-language"] = "python ";
-      myRecordAzOpenAI["custom-copilot-agent"] =
-        "custom-copilot-agent-assistants-api";
-      myRecordAzOpenAI["llm-service"] = "llm-service-openai";
-      myRecordAzOpenAI["openai-key"] = "fake";
+      myRecordAzOpenAI["custom-copilot-agent"] = "custom-copilot-agent-new";
+      myRecordAzOpenAI["llm-service"] = "llm-service-azure-openai";
+      myRecordAzOpenAI["azure-openai-key"] = "fake";
+      myRecordAzOpenAI["azure-openai-deployment-name"] = "fake";
+      myRecordAzOpenAI["azure-openai-endpoint"] = "https://test.com";
       const options = Object.entries(myRecordAzOpenAI)
         .map(([key, value]) => "--" + key + " " + value)
         .join(" ");
@@ -79,13 +80,6 @@ describe("Debug V3 command-and-response template", () => {
         env: process.env,
         timeout: timeout,
       });
-
-      // add extra envs
-      const userFile = path.resolve(projectPath, "env", `.env.local.user`);
-      const OPENAI_ASSISTANT_ID = "OPENAI_ASSISTANT_ID=fake";
-      const KEY = "\n" + OPENAI_ASSISTANT_ID;
-      fs.appendFileSync(userFile, KEY);
-      console.log(`add key ${KEY} to .env.local.user file`);
 
       // provision
       await CliHelper.provisionProject(projectPath, "", "local", {
@@ -131,17 +125,14 @@ describe("Debug V3 command-and-response template", () => {
 
   it(
     "Azure OpenAI happy path: provision and deploy",
-    { testPlanCaseId: 28165244, author: "frankqian@microsoft.com" },
+    { testPlanCaseId: 27689383, author: "frankqian@microsoft.com" },
     async function () {
       // create
       const myRecordAzOpenAI: Record<string, string> = {};
       myRecordAzOpenAI["programming-language"] = "python ";
-      myRecordAzOpenAI["custom-copilot-agent"] =
-        "custom-copilot-agent-assistants-api";
-      myRecordAzOpenAI["llm-service"] = "llm-service-azure-openai";
-      myRecordAzOpenAI["azure-openai-key"] = "fake";
-      myRecordAzOpenAI["azure-openai-deployment-name"] = "fake";
-      myRecordAzOpenAI["azure-openai-endpoint"] = "https://test.com";
+      myRecordAzOpenAI["custom-copilot-agent"] = "custom-copilot-agent-new";
+      myRecordAzOpenAI["llm-service"] = "llm-service-openai";
+      myRecordAzOpenAI["openai-key"] = "fake";
       const options = Object.entries(myRecordAzOpenAI)
         .map(([key, value]) => "--" + key + " " + value)
         .join(" ");
@@ -162,13 +153,6 @@ describe("Debug V3 command-and-response template", () => {
         env: process.env,
         timeout: timeout,
       });
-
-      // add extra envs
-      const userFile = path.resolve(projectPath, "env", `.env.local.user`);
-      const AZURE_OPENAI_ASSISTANT_ID = "AZURE_OPENAI_ASSISTANT_ID=fake";
-      const KEY = "\n" + AZURE_OPENAI_ASSISTANT_ID;
-      fs.appendFileSync(userFile, KEY);
-      console.log(`add key ${KEY} to .env.local.user file`);
 
       // provision
       await CliHelper.provisionProject(projectPath, "", "local", {
