+++ packages/tests/src/e2e/archive/teamsAgent/DebugCustomCopilotBasicBotForPython.tests.bak
@@ -12,13 +12,14 @@ import * as path from "path";
 
 import { it } from "@microsoft/extra-shot-mocha";
 
-import { CliHelper } from "../../commonlib/cliHelper";
+import { CliHelper } from "../../../commonlib/cliHelper";
+import { execAsync } from "../../../utils/commonUtils";
 import {
   cleanUpLocalProject,
   getTestFolder,
   getUniqueAppName,
   readContextMultiEnvV3,
-} from "../commonUtils";
+} from "../../commonUtils";
 import {
   deleteAadAppByClientId,
   deleteBot,
@@ -26,8 +27,7 @@ import {
   getAadAppByClientId,
   getBot,
   getTeamsApp,
-} from "./utility";
-import { execAsync } from "../../utils/commonUtils";
+} from "../../debug/utility";
 
 describe("Debug V3 command-and-response template", () => {
   const testFolder = getTestFolder();
@@ -49,14 +49,12 @@ describe("Debug V3 command-and-response template", () => {
   });
 
   it(
-    "Azure OpenAI happy path: provision and deploy",
-    { testPlanCaseId: 27454153, author: "frankqian@microsoft.com" },
+    "Azure Openai happy path: provision and deploy",
+    { testPlanCaseId: 27178027, author: "frankqian@microsoft.com" },
     async function () {
       // create
       const myRecordAzOpenAI: Record<string, string> = {};
       myRecordAzOpenAI["programming-language"] = "python ";
-      myRecordAzOpenAI["custom-copilot-rag"] =
-        "custom-copilot-rag-azureAISearch";
       myRecordAzOpenAI["llm-service"] = "llm-service-azure-openai";
       myRecordAzOpenAI["azure-openai-key"] = "fake";
       myRecordAzOpenAI["azure-openai-deployment-name"] = "fake";
@@ -67,7 +65,7 @@ describe("Debug V3 command-and-response template", () => {
       await CliHelper.createProjectWithCapability(
         appName,
         testFolder,
-        "custom-copilot-rag" as any,
+        "custom-copilot-basic" as any,
         undefined,
         options
       );
@@ -81,22 +79,6 @@ describe("Debug V3 command-and-response template", () => {
         env: process.env,
         timeout: timeout,
       });
-
-      // add extra envs
-      const userFile = path.resolve(projectPath, "env", `.env.local.user`);
-      const AZURE_OPENAI_EMBEDDING_DEPLOYMENT =
-        "AZURE_OPENAI_EMBEDDING_DEPLOYMENT=fake";
-      const SECRET_AZURE_SEARCH_KEY = "SECRET_AZURE_SEARCH_KEY=fake";
-      const AZURE_SEARCH_ENDPOINT = "AZURE_SEARCH_ENDPOINT=https://test.com";
-      const KEY =
-        "\n" +
-        AZURE_OPENAI_EMBEDDING_DEPLOYMENT +
-        "\n" +
-        SECRET_AZURE_SEARCH_KEY +
-        "\n" +
-        AZURE_SEARCH_ENDPOINT;
-      fs.appendFileSync(userFile, KEY);
-      console.log(`add key ${KEY} to .env.local.user file`);
 
       // provision
       await CliHelper.provisionProject(projectPath, "", "local", {
@@ -141,14 +123,12 @@ describe("Debug V3 command-and-response template", () => {
   );
 
   it(
-    "OpenAI happy path: provision and deploy",
-    { testPlanCaseId: 27454158, author: "frankqian@microsoft.com" },
+    "Openai happy path: provision and deploy",
+    { testPlanCaseId: 27178071, author: "frankqian@microsoft.com" },
     async function () {
       // create
       const myRecordAzOpenAI: Record<string, string> = {};
       myRecordAzOpenAI["programming-language"] = "python ";
-      myRecordAzOpenAI["custom-copilot-rag"] =
-        "custom-copilot-rag-azureAISearch";
       myRecordAzOpenAI["llm-service"] = "llm-service-openai";
       myRecordAzOpenAI["openai-key"] = "fake";
       const options = Object.entries(myRecordAzOpenAI)
@@ -157,7 +137,7 @@ describe("Debug V3 command-and-response template", () => {
       await CliHelper.createProjectWithCapability(
         appName,
         testFolder,
-        "custom-copilot-rag" as any,
+        "custom-copilot-basic" as any,
         undefined,
         options
       );
@@ -171,22 +151,6 @@ describe("Debug V3 command-and-response template", () => {
         env: process.env,
         timeout: timeout,
       });
-
-      // add extra envs
-      const userFile = path.resolve(projectPath, "env", `.env.local.user`);
-      const AZURE_OPENAI_EMBEDDING_DEPLOYMENT =
-        "AZURE_OPENAI_EMBEDDING_DEPLOYMENT=fake";
-      const SECRET_AZURE_SEARCH_KEY = "SECRET_AZURE_SEARCH_KEY=fake";
-      const AZURE_SEARCH_ENDPOINT = "AZURE_SEARCH_ENDPOINT=https://test.com";
-      const KEY =
-        "\n" +
-        AZURE_OPENAI_EMBEDDING_DEPLOYMENT +
-        "\n" +
-        SECRET_AZURE_SEARCH_KEY +
-        "\n" +
-        AZURE_SEARCH_ENDPOINT;
-      fs.appendFileSync(userFile, KEY);
-      console.log(`add key ${KEY} to .env.local.user file`);
 
       // provision
       await CliHelper.provisionProject(projectPath, "", "local", {
