+++ packages/tests/src/e2e/archive/teamsAgent/DebugCustomCopilotRagAiSearchBotForPython.tests.bak
@@ -12,13 +12,14 @@ import * as path from "path";
 
 import { it } from "@microsoft/extra-shot-mocha";
 
-import { CliHelper } from "../../commonlib/cliHelper";
+import { CliHelper } from "../../../commonlib/cliHelper";
+import { execAsync } from "../../../utils/commonUtils";
 import {
   cleanUpLocalProject,
   getTestFolder,
   getUniqueAppName,
   readContextMultiEnvV3,
-} from "../commonUtils";
+} from "../../commonUtils";
 import {
   deleteAadAppByClientId,
   deleteBot,
@@ -26,8 +27,7 @@ import {
   getAadAppByClientId,
   getBot,
   getTeamsApp,
-} from "./utility";
-import { execAsync } from "../../utils/commonUtils";
+} from "../../debug/utility";
 
 describe("Debug V3 command-and-response template", () => {
   const testFolder = getTestFolder();
@@ -50,12 +50,13 @@ describe("Debug V3 command-and-response template", () => {
 
   it(
     "Azure OpenAI happy path: provision and deploy",
-    { testPlanCaseId: 27551381, author: "frankqian@microsoft.com" },
+    { testPlanCaseId: 27454153, author: "frankqian@microsoft.com" },
     async function () {
       // create
       const myRecordAzOpenAI: Record<string, string> = {};
       myRecordAzOpenAI["programming-language"] = "python ";
-      myRecordAzOpenAI["custom-copilot-rag"] = "custom-copilot-rag-customize";
+      myRecordAzOpenAI["custom-copilot-rag"] =
+        "custom-copilot-rag-azureAISearch";
       myRecordAzOpenAI["llm-service"] = "llm-service-azure-openai";
       myRecordAzOpenAI["azure-openai-key"] = "fake";
       myRecordAzOpenAI["azure-openai-deployment-name"] = "fake";
@@ -80,6 +81,22 @@ describe("Debug V3 command-and-response template", () => {
         env: process.env,
         timeout: timeout,
       });
+
+      // add extra envs
+      const userFile = path.resolve(projectPath, "env", `.env.local.user`);
+      const AZURE_OPENAI_EMBEDDING_DEPLOYMENT =
+        "AZURE_OPENAI_EMBEDDING_DEPLOYMENT=fake";
+      const SECRET_AZURE_SEARCH_KEY = "SECRET_AZURE_SEARCH_KEY=fake";
+      const AZURE_SEARCH_ENDPOINT = "AZURE_SEARCH_ENDPOINT=https://test.com";
+      const KEY =
+        "\n" +
+        AZURE_OPENAI_EMBEDDING_DEPLOYMENT +
+        "\n" +
+        SECRET_AZURE_SEARCH_KEY +
+        "\n" +
+        AZURE_SEARCH_ENDPOINT;
+      fs.appendFileSync(userFile, KEY);
+      console.log(`add key ${KEY} to .env.local.user file`);
 
       // provision
       await CliHelper.provisionProject(projectPath, "", "local", {
@@ -124,13 +141,14 @@ describe("Debug V3 command-and-response template", () => {
   );
 
   it(
-    "Azure OpenAI happy path: provision and deploy",
-    { testPlanCaseId: 27551384, author: "frankqian@microsoft.com" },
+    "OpenAI happy path: provision and deploy",
+    { testPlanCaseId: 27454158, author: "frankqian@microsoft.com" },
     async function () {
       // create
       const myRecordAzOpenAI: Record<string, string> = {};
       myRecordAzOpenAI["programming-language"] = "python ";
-      myRecordAzOpenAI["custom-copilot-rag"] = "custom-copilot-rag-customize";
+      myRecordAzOpenAI["custom-copilot-rag"] =
+        "custom-copilot-rag-azureAISearch";
       myRecordAzOpenAI["llm-service"] = "llm-service-openai";
       myRecordAzOpenAI["openai-key"] = "fake";
       const options = Object.entries(myRecordAzOpenAI)
@@ -153,6 +171,22 @@ describe("Debug V3 command-and-response template", () => {
         env: process.env,
         timeout: timeout,
       });
+
+      // add extra envs
+      const userFile = path.resolve(projectPath, "env", `.env.local.user`);
+      const AZURE_OPENAI_EMBEDDING_DEPLOYMENT =
+        "AZURE_OPENAI_EMBEDDING_DEPLOYMENT=fake";
+      const SECRET_AZURE_SEARCH_KEY = "SECRET_AZURE_SEARCH_KEY=fake";
+      const AZURE_SEARCH_ENDPOINT = "AZURE_SEARCH_ENDPOINT=https://test.com";
+      const KEY =
+        "\n" +
+        AZURE_OPENAI_EMBEDDING_DEPLOYMENT +
+        "\n" +
+        SECRET_AZURE_SEARCH_KEY +
+        "\n" +
+        AZURE_SEARCH_ENDPOINT;
+      fs.appendFileSync(userFile, KEY);
+      console.log(`add key ${KEY} to .env.local.user file`);
 
       // provision
       await CliHelper.provisionProject(projectPath, "", "local", {
